Code review rounds 7-9: fix 27 bugs across 12 files, add multi-marker profile
Critical fixes:
- run_segmentation.py: UID int() → int(round()) to match tile_processing.py
- run_segmentation.py: auto-enable --all-channels for cell + cellpose_input_channels
- flat_field.py: coarse grid float64→float32 (halves zoom() memory from 174GB to 87GB)
- classify_markers.py: Otsu crash guard on zero-variance, empty values guard
- spatial_cell_analysis.py: nx.diameter capped at 1000 (double-BFS for larger)
- run_pipeline.sh: mixed-method validation, respect enabled:false, sbatch in output dir

High fixes:
- flat_field.py: dtype-aware clip/cast (supports uint8/uint16, not hardcoded)
- cell.py: removed duplicate SAM2 set_image (saves ~1h GPU per job)
- cell.py: grayscale fallback refactored to _grayscale_cellpose_eval with _safe_to_uint8
- run_segmentation.py: set_channel_data for ALL channels after flat-field
- spatial_cell_analysis.py: vectorized edge distances, degenerate Delaunay guard
- mixins.py + feature_extraction.py + base.py: .astype(float) → .astype(np.float32)

New features:
- classify_markers.py: combined marker_profile field for multi-marker configs
- vessel config: spatial_viewer with marker_profile grouping (4 SMA/CD31 combos)
- msln config: spatial_viewer added, Pdgfra ch3 dropped (stain failed)

+++ segmentation/detection/strategies/base.py
@@ -219,9 +219,9 @@ class DetectionStrategy(ABC):
 
         # Get intensity image
         if tile.ndim == 3:
-            gray = np.mean(tile[:, :, :3], axis=2)
+            gray = np.mean(tile[:, :, :3].astype(np.float32), axis=2)
         else:
-            gray = tile.astype(float)
+            gray = tile.astype(np.float32)
 
         # Use largest connected component for shape features (solidity, eccentricity)
         # to avoid picking a small fragment when mask has disconnected components
@@ -648,7 +648,7 @@ class DetectionStrategy(ABC):
         if p_high_val <= p_low_val:
             return np.zeros_like(image, dtype=np.uint8)
 
-        normalized = (image.astype(float) - p_low_val) / (p_high_val - p_low_val)
+        normalized = (image.astype(np.float32) - p_low_val) / (p_high_val - p_low_val)
         normalized = np.clip(normalized * 255, 0, 255).astype(np.uint8)
 
         return normalized
